moved ValueCounts class in the Metrics folder
