+++ src/explorer/Components/Datetime/LinearTimeBuckets.cs
@@ -7,6 +7,7 @@ namespace Explorer.Components
 
     using Explorer.Common;
     using Explorer.Common.Utils;
+    using Explorer.Metrics;
     using Explorer.Queries;
 
     public class LinearTimeBuckets : ExplorerComponent<LinearTimeBuckets.Result>, PublisherComponent
